Usage Stats: Split domain & service packages
