+++ pkg/infra/usagestats/service/usage_stats_test.go
@@ -1,4 +1,4 @@
-package usagestats
+package service
 
 import (
 	"bytes"
@@ -16,6 +16,7 @@ import (
 	"github.com/grafana/grafana/pkg/components/simplejson"
 	"github.com/grafana/grafana/pkg/infra/kvstore"
 	"github.com/grafana/grafana/pkg/infra/log"
+	"github.com/grafana/grafana/pkg/infra/usagestats"
 	"github.com/grafana/grafana/pkg/models"
 	"github.com/grafana/grafana/pkg/plugins"
 	"github.com/grafana/grafana/pkg/plugins/manager"
@@ -29,10 +30,10 @@ import (
 
 // This is to ensure that the interface contract is held by the implementation
 func Test_InterfaceContractValidity(t *testing.T) {
-	newUsageStats := func() UsageStats {
-		return &UsageStatsService{}
+	newUsageStats := func() usagestats.Service {
+		return &UsageStats{}
 	}
-	v, ok := newUsageStats().(*UsageStatsService)
+	v, ok := newUsageStats().(*UsageStats)
 
 	assert.NotNil(t, v)
 	assert.True(t, ok)
@@ -221,7 +222,7 @@ func TestMetrics(t *testing.T) {
 				sendUsageStats = origSendUsageStats
 			})
 			statsSent := false
-			sendUsageStats = func(uss *UsageStatsService, b *bytes.Buffer) {
+			sendUsageStats = func(uss *UsageStats, b *bytes.Buffer) {
 				statsSent = true
 			}
 
@@ -592,7 +593,7 @@ func (pm *fakePluginManager) PanelCount() int {
 	return len(pm.panels)
 }
 
-func setupSomeDataSourcePlugins(t *testing.T, uss *UsageStatsService) {
+func setupSomeDataSourcePlugins(t *testing.T, uss *UsageStats) {
 	t.Helper()
 
 	uss.PluginManager = &fakePluginManager{
@@ -635,17 +636,17 @@ type httpResp struct {
 	err            error
 }
 
-func createService(t *testing.T, cfg setting.Cfg) *UsageStatsService {
+func createService(t *testing.T, cfg setting.Cfg) *UsageStats {
 	t.Helper()
 
 	sqlStore := sqlstore.InitTestDB(t)
 
-	return &UsageStatsService{
+	return &UsageStats{
 		Bus:                bus.New(),
 		Cfg:                &cfg,
 		SQLStore:           sqlStore,
 		AlertingUsageStats: &alertingUsageMock{},
-		externalMetrics:    make([]MetricsFunc, 0),
+		externalMetrics:    make([]usagestats.MetricsFunc, 0),
 		PluginManager:      &fakePluginManager{},
 		grafanaLive:        newTestLive(t),
 		kvStore:            kvstore.WithNamespace(kvstore.ProvideService(sqlStore), 0, "infra.usagestats"),
